switch

+++ user/uthread.c
@@ -80,7 +80,7 @@ thread_schedule(void)
      * Invoke thread_switch to switch from t to next_thread:
      * thread_switch(??, ??);
      */
-    thread_switch(&(t->context), &(current_thread->context));
+    thread_switch((uint64)(&(t->context)), (uint64)(&(current_thread->context)));
   } else
     next_thread = 0;
 }
